requirements.txt
absl-py==2.3.1
astunparse==1.6.3
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
cloudpickle==3.1.1
decorator==5.2.1
dm-tree==0.1.9
filelock==3.13.1
flatbuffers==25.2.10
fsspec==2024.6.1
gast==0.4.0
google-auth==2.40.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
Jinja2==3.1.4
keras==2.15.0
libclang==18.1.1
Markdown==3.8.2
MarkupSafe==2.1.5
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.3
numpy==1.24.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.3.1
onnx==1.14.1
onnx-tf==1.10.0
opt_einsum==3.4.0
packaging==25.0
pillow==11.0.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyYAML==6.0.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.10.1
six==1.17.0
sympy==1.13.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.22.0
termcolor==3.1.0
tflite-runtime==2.13.0
# torch==2.1.0+cpu
# torchaudio==2.1.0+cpu
torchinfo==1.8.0
# torchvision==0.16.0+cpu
tqdm==4.66.4
triton==3.2.0
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.1
